refactor(agents): separate child execution boundaries
